+++ cpp04/ex03/includes/ICharacter.hpp
@@ -6,7 +6,7 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/01 13:01:08 by mfeldman          #+#    #+#             */
-/*   Updated: 2024/02/05 12:29:12 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/02/07 15:11:23 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 #include "AMateria.hpp"
 #include <string>
 #include <iostream>
+#include <stdint.h>
 
 class AMateria;
 
@@ -22,7 +23,7 @@ class ICharacter
 	public:
 		virtual	~ICharacter() {}
 		virtual void				equip(AMateria *m) = 0;
-		virtual void				unequip(int idx) = 0;
-		virtual void				use(int idx, ICharacter &target) = 0;
+		virtual void				unequip(uint8_t idx) = 0;
+		virtual void				use(uint8_t idx, ICharacter &target) = 0;
 		virtual std::string const&	get_name() const = 0;
 };
